requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
huggingface_hub
langchain
langchain-huggingface
pydantic
python-dotenv